chore: rename globalState folder to store
