+++ src/Components/CountryPicker/CountryPicker.js
@@ -1,6 +1,6 @@
 import { useTranslation } from "react-i18next";
 import { useDispatch } from "react-redux";
-import { setSelectedCountry } from "../../globalState/slices/borderCrossingData";
+import { setSelectedCountry } from "../../store/slices/borderCrossingData";
 
 const CountryPicker = ({ availableCountries, selectedCountry }) => {
   const dispatch = useDispatch();
