+++ src/store/configureStore.js
@@ -2,7 +2,7 @@ import { configureStore } from "@reduxjs/toolkit";
 import currentPage from "./slices/currentPage";
 import borderCrossingData from "./slices/borderCrossingData";
 
-export const store = configureStore({
+export default configureStore({
   reducer: {
     currentPage,
     borderCrossingData,
